+++ test.c
@@ -27,13 +27,13 @@ int main(int   argc,
     if(zinit(argv, (Flag[]){num, str}, 2))
         return 0;
     
-    char **num_list = flag_value(argv, &num);
-    for (int i = 0; i < flag_count(argv, &num); i++)
+    char **num_list = flag_value(argv, num);
+    for (int i = 0; i < flag_count(argv, num); i++)
         printf("Number %d: %d\n", i, atoi(num_list[i]));
     free(num_list);
     
-    char **str_list = flag_value(argv, &str);
-    for (int i = 0; i < flag_count(argv, &str); i++)
+    char **str_list = flag_value(argv, str);
+    for (int i = 0; i < flag_count(argv, str); i++)
         printf("String %d: %s\n", i, str_list[i]);
     free(str_list);